third_party/blink/web_tests/svg/dynamic-updates/SVGFilterElement-svgdom-width-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'width' property of the SVGFilterElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS filterElement.width.baseVal.value is 100
PASS filterElement.width.baseVal.value is 200
PASS successfullyParsed is true

TEST COMPLETE
